Adds workarounds for not being able to place sentries as client
This bug can happen when the server disagrees with the client about the sentry location being valid, in which case the client will forever wait for the server to spawn the sentry and no longer be able to place any sentry guns.
The server-side workaround adjusts the sentry attachment code to be consistent with the sentry placement preview and the client-side workaround cancels waiting for the server response after five seconds of no sentry being spawned.